# Speed up "VMs & Instances I Own" filter in /vm_or_template/explorer