Purpose or Intent
-----------------
Speeds up "VMs & Instances I Own" index and ajax calls in the `/vm_or_template` (Services -> Workload) section of the application.

How
---
This adds `virtual_attribute` on to the `ownership_mixin`, which is used by a few classes and makes it so that MiqExpressions will evaluate filter in SQL instead of pulling down all matching records to then iterate over and filter in Ruby (causes a Massive N+1 on large datasets).  This requires a join for the `virtual_attribute` to function, so the `with_evm_owner` scope does that with out pulling in any extra data.  This is added in the controller in `set_named_scope`, and is eventually used in `MiqReport#paged_view_search` with an `MiqExpression`.


Metrics
-------
The data set for these metrics used a DB with approximately 20k `Vm` records.

**Before**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 92890ms (Views: 2.9ms | ActiveRecord: 11462.1ms)

|       ms | bytes |    objects | queries | query (ms) |    rows |
|     ---: |  ---: |       ---: |    ---: |       ---: |    ---: |
| 92,346.1 |   N/A | 54,281,990 |     188 |   70,430.0 | 349,165 |



**After**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 748ms (Views: 3.5ms | ActiveRecord: 150.8ms)

|       ms | bytes |    objects | queries | query (ms) |    rows |
|     ---: |  ---: |       ---: |    ---: |       ---: |    ---: |
|    639.2 |   N/A |  1,100,258 |      26 |      330.2 |     365 |


Additional Notes
----------------

This is based off of the changes in https://github.com/ManageIQ/manageiq/pull/10290 and so that PR is required before this one can be merged.  Many of the gotcha's and such are also valid in this PR as well.

A easier to view diff can be found here:  https://github.com/NickLaMuro/manageiq/compare/ownership_mixin_owned_by_ldap_sql...ownership_mixin_owned_by_user_sql