@chessbyte 2016-08-23
@NickLaMuro should this be closed in favor of #10685 ?