@NickLaMuro 2016-08-23
@chessbyte well, technically @kbrock was working on something to that PR specifically for this query, not sure if it is up yet though.

Regardless, this is not how we are going to move forward with improving this one.